requirements.txt
streamlit
transformers
pillow
requests
torch
tensorflow
